Added a checkout form and basic form validation
Added Checkout.js for a user to enter name, street, postal code, and city if user wants to wants to place order. Implemented basic validation for the form inputs.

There is an issue with styling where adding more than 2 unique items in cart shows to scroll bars in my computer.